feat(NaturePath): refactor and split
