+++ NaturePathfinder/Models/Area.cs
@@ -1,0 +1,7 @@
+namespace ExotischNederland.Models
+{
+    public class Area
+    {
+        
+    }
+}
