+++ NaturePathfinder/Models/Route.cs
@@ -1,0 +1,7 @@
+namespace ExotischNederland.Models
+{
+    public class Route
+    {
+        
+    }
+}
