+++ NaturePathfinder/NaturePathfinder.csproj
@@ -40,14 +40,22 @@
         <Reference Include="System.Xml"/>
     </ItemGroup>
     <ItemGroup>
+        <Compile Include="Helpers\DistanceCalculator.cs" />
+        <Compile Include="Models\Area.cs" />
+        <Compile Include="Models\Route.cs" />
         <Compile Include="Program.cs"/>
         <Compile Include="Properties\AssemblyInfo.cs"/>
+        <Compile Include="Services\Dijkstra.cs" />
+        <Compile Include="Services\Graph.cs" />
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\ExotischNederland\ExotischNederland.csproj">
         <Project>{8ac3c931-f28f-434e-ab52-776d9a9e73e3}</Project>
         <Name>ExotischNederland</Name>
       </ProjectReference>
+    </ItemGroup>
+    <ItemGroup>
+      <Content Include="README.md" />
     </ItemGroup>
     <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets"/>
     <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
